requirements.txt
requests==2.25.1
gate_api==4.22.2
PyYAML==6.0
pika==1.2.*
pytest==6.2.*
pylint==2.11.*
pytest-asyncio==0.16.*